+++ t2eldencomptest/src/main/java/com/elden/t2eldencomptest/activities/friend_details/FriendDetailsActivity.java
@@ -1,12 +1,8 @@
-package com.elden.t2eldencomptest.activities.friend_list;
+package com.elden.t2eldencomptest.activities.friend_details;
 
 import android.databinding.DataBindingUtil;
 import android.os.Bundle;
-import android.support.design.widget.FloatingActionButton;
-import android.support.design.widget.Snackbar;
 import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.CardView;
-import android.support.v7.widget.Toolbar;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
